final adjustments to firebase operations and related functions
