+++ src/firebase/operations/delete.js
@@ -1,11 +1,7 @@
 import { deleteDoc as firebaseDeleteDoc } from 'firebase/firestore';
 
 export const deleteDoc = async (documentRef) => {
-  try {
-    await firebaseDeleteDoc(documentRef);
+  await firebaseDeleteDoc(documentRef);
 
-    return documentRef.id;
-  } catch (error) {
-    return error;
-  }
+  return documentRef.id;
 };
